+++ CHANGELOG.md
@@ -4,7 +4,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 
 ## [{version}] - {release_date}
 ### Added
-- 
+- Funders now have an "additional comments" field
 
 ### Changed
 -